# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "MealTracker API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/users/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/meals/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      },
      "post": {
        "tags": [
          "Meals"
        ],
        "description": "Creates a new meal.",
        "parameters": [
          {
            "name": "body",
            "typeId": "body",
            "calories": "body",
            "protein": "body",
            "schema": {
              "$ref": "#/definitions/Meal"
            },
            "in": "query"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "MealName": {
                  "example": "any"
                },
                "MealType": {
                  "example": "any"
                },
                "Calories": {
                  "example": "any"
                },
                "Protein": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "MealId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/profile/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    }
  },
  "definitions": {
    "Meal": {
      "type": "object",
      "properties": {
        "Name": {
          "type": "string",
          "example": "Pasta bolognese"
        },
        "TypeId": {
          "type": "number",
          "example": 3
        },
        "Calories": {
          "type": "number",
          "example": 540
        },
        "Protein": {
          "type": "number",
          "example": 40
        }
      },
      "required": [
        "Name",
        "TypeId",
        "Calories",
        "Protein"
      ]
    }
  }
}